+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Support log pattern in discover summary ([#550](https://github.com/opensearch-project/dashboards-assistant/pull/550))
 
 ### Bug Fixes
+- Hide AI features for AnalyticEngine data sources ([#689](https://github.com/opensearch-project/dashboards-assistant/pull/689))
 - detect serverless data source ([#627](https://github.com/opensearch-project/dashboards-assistant/pull/627))
 - Fix capability services access settings before login and show dialog ([#628](https://github.com/opensearch-project/dashboards-assistant/pull/628))
 - disable input panel and allow user click new conversation in error mode ([#639](https://github.com/opensearch-project/dashboards-assistant/pull/639))